Improve gem structure
- Use require instead of require_relative when possible
- Move some logic in `lib/vimrunner.r` to `lib/vimrunner/vimrunner.rb`
